% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cleaned_holiday_movie_genres}
\alias{cleaned_holiday_movie_genres}
\title{Holiday Movie Genres Dataset}
\format{
A data frame with x rows and 2 variables:
\describe{
\item{tconst}{Unique identifier for the movie (character)}
\item{genres}{Genres associated with the movie (character)}
}
}
\source{
The data was obtained from the TidyTuesday project. TidyTuesday is a weekly data project.
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2023/2023-12-12}
}
\usage{
cleaned_holiday_movie_genres
}
\description{
A dataset containing genres associated with holiday movies, where each movie may have multiple genres listed in separate rows. Irrelevant genres like "Music", "Talk-Show", "News" have been removed.
}
\keyword{datasets}
