% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holiday_movies}
\alias{holiday_movies}
\title{Holiday Movies Dataset}
\format{
A data frame with x rows and y variables:
\describe{
\item{tconst}{Unique identifier for the movie (character)}
\item{primary_title}{Popular title of the movie (character)}
\item{year}{Year the movie was released (numeric)}
\item{genres}{Comma-separated list of genres (character)}
\item{christmas}{1 if the movie relates to Christmas, 0 otherwise (integer)}
\item{hanukkah}{1 if the movie relates to Hanukkah, 0 otherwise (integer)}
\item{kwanzaa}{1 if the movie relates to Kwanzaa, 0 otherwise (integer)}
\item{average_rating}{IMDb average rating (numeric)}
\item{num_votes}{Number of votes received on IMDb (numeric)}
}
}
\source{
The data was obtained from the TidyTuesday project. TidyTuesday is a weekly data project.
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2023/2023-12-12}
}
\usage{
holiday_movies
}
\description{
A dataset containing holiday-related movies from 1929 to 2023. The data includes movie titles, genres, IMDb ratings, and holiday types (Christmas, Hanukkah, Kwanzaa).
}
\keyword{datasets}
